Improve structure size/packing of gcfg_number_t
The digit counter doesn't need to be 16 bits, we can have at
most 10 decimal digits with the 32 bit fraction field.

Making it and the flags 8 bits, and re-arranging the struct
a bit gives us a slightly improved layout.

+++ lib/parse_number.c
@@ -36,7 +36,7 @@ const char *gcfg_parse_number(gcfg_file_t *f, const char *in,
 		if (in == NULL)
 			return NULL;
 
-		num->fraction_digits = (uint16_t)i;
+		num->fraction_digits = (uint8_t)i;
 		num->fraction = (uint32_t)temp;
 	} else {
 		num->fraction_digits = 0;
